#362 Visual Cleanup
