+++ src/frontend/src/pages/TeamsPage/Teams.tsx
@@ -1,3 +1,8 @@
+/*
+ * This file is part of NER's FinishLine and licensed under GNU AGPLv3.
+ * See the LICENSE file in the repository root folder for details.
+ */
+
 import { Route, Switch } from 'react-router-dom';
 import { routes } from '../../utils/routes';
 import TeamsPage from './TeamsPage';
